Refactored getProductRating()
