+++ src/main/java/be/syntra/devshop/DevshopBack/services/ProductService.java
@@ -22,4 +22,6 @@ public interface ProductService {
     List<Product> findAllByNameContainingIgnoreCaseAndArchivedFalse(String searchRequest);
 
     Product findById(Long id);
+
+    Double getProductRating(Long productId);
 }
