+++ src/main/java/be/syntra/devshop/DevshopBack/services/StarRatingService.java
@@ -6,7 +6,5 @@ import java.util.Set;
 
 public interface StarRatingService {
 
-    Double getProductRating();
-
     Set<StarRating> findAll();
 }
